% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootTest.R
\name{bootTest}
\alias{bootTest}
\title{Bootstrap test for underlying non-normality in ordinal data.}
\usage{
bootTest(my.data, B = 1000)
}
\arguments{
\item{my.data}{a dataset containing ordinal data.}

\item{B}{number of bootstrap samples.}
}
\value{
p-value associated with the underlying normality hypothesis.
}
\description{
Bootstrap test for underlying non-normality in ordinal data.
}
